refactor:move left sidebar to components
